requirements.txt
flask==2.2.3
pandas==1.5.3
openpyxl==3.1.2
Werkzeug==2.2.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
pyjwt==2.6.0 